requirements.txt
python-emailahoy
djangorestframework
django
itchat
email
